Move ChequeCommand to command package and refactor Bank
